refactor: children view merged in relations view
